+++ commands.js
@@ -3,6 +3,7 @@ import { resetRareCatchesTracker } from './features/overlays/rareCatchesTracker'
 import { resetCrimsonIsleTracker } from "./features/overlays/crimsonIsleTracker";
 import { resetJerryWorkshopTracker } from "./features/overlays/jerryWorkshopTracker";
 import { resetWormMembraneProfitTracker } from "./features/overlays/wormMembraneProfitTracker";
+import { resetMagmaCoreProfitTracker } from "./features/overlays/magmaCoreProfitTracker";
 
 register("command", (...args) => {
     settings.openGUI();
@@ -32,3 +33,9 @@ register("command", (...args) => {
     resetWormMembraneProfitTracker(!!isConfirmed);
     return;
 }).setName("feeshResetWormProfit");
+
+register("command", (...args) => {
+    const isConfirmed = args[0] && args[0] === "noconfirm";
+    resetMagmaCoreProfitTracker(!!isConfirmed);
+    return;
+}).setName("feeshResetMagmaCoreProfit");