+++ docs/CHANGELOG.md
@@ -1,11 +1,19 @@
 # Releases
 
+## v1.13.0
+
+- Magma Core profit tracker overlay.
+- Added alert on Spirit Mask used.
+- Added small delay for Thunder Bottle Full alert (so it should appear after Thunder spawn alert).
+
+## v1.12.0
+
 - Magic find shown in the drop messages sent to the party chat (toggleable) - [!] requres your party to update the module /ct load [!].
 - Alert on Music Rune I caught.
 - Changed Radiactive Vial drop sound to more satisfying.
 - Alert on Chum bucket automatically picked up when you moved away from it.
 - Added button to pause worm profit tracker (it still pauses automatically after 1 minute of inactivity as before).
-- Set minimal amount of sea creatures cap to 5 (settings)
+- Set minimal amount of sea creatures cap to 5 in the settings.
 - Random fixes in different areas.
 
 ## v1.11.0